+++ packages/cli/package.json
@@ -1,6 +1,6 @@
 {
   "name": "fossyl",
-  "version": "0.18.4",
+  "version": "0.18.5",
   "type": "module",
   "description": "CLI for scaffolding fossyl projects",
   "bin": {
