+++ packages/core/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@fossyl/core",
-  "version": "0.18.4",
+  "version": "0.18.5",
   "repository": {
     "type": "git",
     "url": "https://github.com/YoyoSaur/fossyl.git",
