+++ packages/zod/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@fossyl/zod",
-  "version": "0.18.4",
+  "version": "0.18.5",
   "description": "Zod validation adapter for fossyl",
   "main": "./dist/index.js",
   "module": "./dist/index.mjs",
